+++ script/writeoff_tmp.pl
@@ -6,7 +6,6 @@ use 5.014;
 BEGIN { push @INC, 'lib'; }
 
 use WriteOff::Schema;
-use DateTime;
 my $schema = WriteOff::Schema->connect('dbi:SQLite:WriteOff.db');
 
-say $_->args for $schema->resultset('Schedule')->all
+say $schema->resultset('Image')->find(25)->alt;