+++ Changes.md
@@ -9,6 +9,7 @@ v0.28
 - Made judge votes listed in the Controller::Event::results
   - Utilised session data to determine what record view the user is in
 - Added action Event::overview for permalinks to event overviews
+- Added clause about vote doctoring to rules
 
 v0.27
 =====